Add class IndentedCommandSenderMessenger and pull shared code with IndentedLogger to abstract class IndentedStringHandler
